+++ Demo/API/Assets/WX-WASM-SDK-V2/Editor/WXExtDef.cs
@@ -66,6 +66,11 @@ namespace WeChatWASM
 #else
             WXExtEnvDef.SETDEF("UNITY_2022", false);
 #endif
+#if UNITY_2022_2_OR_NEWER
+            WXExtEnvDef.SETDEF("UNITY_2022_2_OR_NEWER", true);
+#else
+            WXExtEnvDef.SETDEF("UNITY_2022_2_OR_NEWER", false);
+#endif
 #if UNITY_INSTANTGAME
             WXExtEnvDef.SETDEF("UNITY_INSTANTGAME", true);
 #else